cwe-tests/python/django/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
requests>=2.28
